+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ## [2.0.0+2] - 
+* Reduce final .apk app size when using this lib
 * Fixed `ValueDecoration` when all items were 0
 * Add value label generator in `ValueDecoration`
 * Add show the line for value to `HorizontalDecoration` for more precise control of when to show horizontal line
+* Fix gradient color on `SparkLineDecoration`
 * Fix fixed horizontal decoration values getting clipped
 
 ## [2.0.0+1] - 01.03.2022.